port: avoid truncating link speed values above 64gb/s
Ports with higher speeds (100gb/s and 400gb/s) are not represented in
the API. The field is only a 16bit wide integer. These values are
truncated.

Change the field to use 32 bits as it is in the DPDK API.

Fixes: 5a25e184cb1d ("port: expose link_speed to the api")

+++ modules/infra/cli/port.c
@@ -21,7 +21,7 @@ static void port_show(const struct gr_api_client *c, const struct gr_iface *ifac
 	printf("devargs: %s\n", port->devargs);
 	printf("driver:  %s\n", port->driver_name);
 	printf("mac: " ETH_F "\n", &port->mac);
-	if (port->link_speed == UINT16_MAX)
+	if (port->link_speed == UINT32_MAX)
 		printf("speed: unknown\n");
 	else
 		printf("speed: %u Mb/s\n", port->link_speed);
